Split UnwinnabilityMaterial into BitboardPosition and StaticPosition classes
Resolves Checkstyle OverloadMethodsDeclarationOrder warnings firing across
the StaticPosition/BitboardPosition method pairs. Splits the file into:

- UnwinnabilityMaterialBitboard (new) — BitboardPosition variants only;
  the production class consumed by UnwinnableQuickAnalyzer,
  FindHelpmateExhaust, and GoingToCorner.
- UnwinnabilityMaterial — StaticPosition variants only; the differential-
  test oracle consumed only by TestUnwinnabilityMaterialBitboard.

No src/main caller used the StaticPosition variants anymore — they had
all been ported in the switchover release. Splitting now pre-positions
UnwinnabilityMaterial.java for the Phase 6 relocation: when StaticPosition
moves to src/test/, this class moves with it as a single `git mv`, no
surgical method extraction.
