+++ src/main/java/com/dlb/chess/unwinnability/UnwinnabilityMaterial.java
@@ -1,6 +1,5 @@
 package com.dlb.chess.unwinnability;
 
-import com.dlb.chess.bitboard.BitboardPosition;
 import com.dlb.chess.board.StaticPosition;
 import com.dlb.chess.board.enums.Piece;
 import com.dlb.chess.board.enums.PieceType;
@@ -10,9 +9,11 @@ import com.dlb.chess.board.enums.SquareType;
 import com.dlb.chess.common.constants.EnumConstants;
 
 /**
- * Internal material predicates used by the unwinnability/helpmate analysis. Not part of the public API: the library's
- * public surface is the {@link UnwinnableFullAnalyzer} / {@link UnwinnableQuickAnalyzer} entry points, not the
- * per-square material arithmetic.
+ * StaticPosition-backed material predicates used by the unwinnability/helpmate analysis. Reference implementations
+ * of the same predicates that {@link UnwinnabilityMaterialBitboard} computes from a {@link com.dlb.chess.bitboard
+ * .BitboardPosition}. This class is the differential-test oracle — production callers all consume the bitboard
+ * sibling. When Phase 6 of the switchover lands and the StaticPosition subtree relocates to {@code src/test/}, this
+ * class moves with it as a single {@code git mv}.
  */
 abstract class UnwinnabilityMaterial implements EnumConstants {
 
@@ -125,105 +126,7 @@ abstract class UnwinnabilityMaterial implements EnumConstants {
         && countPieces(side, staticPosition, PieceType.PAWN) == 0;
   }
 
-  // ===========================================================================================
-  // Bitboard variants. Same predicates, BitboardPosition input. Used by callers being ported off
-  // the StaticPosition surface as part of the switchover release.
-  // ===========================================================================================
-
-  // Square-colour masks: A1 is dark, B1 is light, alternating. Bit i is light iff (file_i + rank_i) is odd.
-  private static final long LIGHT_SQUARES = 0x55AA55AA55AA55AAL;
-  private static final long DARK_SQUARES = ~LIGHT_SQUARES;
-
-  static boolean calculateHasRook(BitboardPosition bitboardPosition) {
-    return (bitboardPosition.whiteRooks() | bitboardPosition.blackRooks()) != 0L;
-  }
-
-  static boolean calculateHasRook(Side side, BitboardPosition bitboardPosition) {
-    return (side == Side.WHITE ? bitboardPosition.whiteRooks() : bitboardPosition.blackRooks()) != 0L;
-  }
-
-  static boolean calculateHasKnight(BitboardPosition bitboardPosition) {
-    return (bitboardPosition.whiteKnights() | bitboardPosition.blackKnights()) != 0L;
-  }
-
-  static boolean calculateHasKnight(Side side, BitboardPosition bitboardPosition) {
-    return (side == Side.WHITE ? bitboardPosition.whiteKnights() : bitboardPosition.blackKnights()) != 0L;
-  }
-
-  static boolean calculateHasQueen(BitboardPosition bitboardPosition) {
-    return (bitboardPosition.whiteQueens() | bitboardPosition.blackQueens()) != 0L;
-  }
-
-  static boolean calculateHasQueen(Side side, BitboardPosition bitboardPosition) {
-    return (side == Side.WHITE ? bitboardPosition.whiteQueens() : bitboardPosition.blackQueens()) != 0L;
-  }
-
-  static boolean calculateHasNoRooks(Side side, BitboardPosition bitboardPosition) {
-    return !calculateHasRook(side, bitboardPosition);
-  }
-
-  static boolean calculateHasNoKnights(Side side, BitboardPosition bitboardPosition) {
-    return !calculateHasKnight(side, bitboardPosition);
-  }
-
-  static boolean calculateHasNoBishops(Side side, BitboardPosition bitboardPosition) {
-    return (side == Side.WHITE ? bitboardPosition.whiteBishops() : bitboardPosition.blackBishops()) == 0L;
-  }
-
-  static boolean calculateHasNoBishops(Side side, BitboardPosition bitboardPosition, SquareType squareType) {
-    return countBishops(side, bitboardPosition, squareType) == 0;
-  }
-
-  static boolean calculateHasNoPawns(Side side, BitboardPosition bitboardPosition) {
-    return (side == Side.WHITE ? bitboardPosition.whitePawns() : bitboardPosition.blackPawns()) == 0L;
-  }
-
-  static boolean calculateHasLightSquareBishops(Side side, BitboardPosition bitboardPosition) {
-    final long bishops = side == Side.WHITE ? bitboardPosition.whiteBishops() : bitboardPosition.blackBishops();
-    return (bishops & LIGHT_SQUARES) != 0L;
-  }
-
-  static boolean calculateHasDarkSquareBishops(Side side, BitboardPosition bitboardPosition) {
-    final long bishops = side == Side.WHITE ? bitboardPosition.whiteBishops() : bitboardPosition.blackBishops();
-    return (bishops & DARK_SQUARES) != 0L;
-  }
-
-  static boolean calculateHasKingOnly(Side side, BitboardPosition bitboardPosition) {
-    final long sideOccupancy = bitboardPosition.occupied(side);
-    final long sideKings = side == Side.WHITE ? bitboardPosition.whiteKings() : bitboardPosition.blackKings();
-    return sideOccupancy == sideKings && Long.bitCount(sideKings) == 1;
-  }
-
-  static boolean calculateHasKingAndKnightOnly(Side side, BitboardPosition bitboardPosition) {
-    final long sideKings = side == Side.WHITE ? bitboardPosition.whiteKings() : bitboardPosition.blackKings();
-    final long sideKnights = side == Side.WHITE ? bitboardPosition.whiteKnights() : bitboardPosition.blackKnights();
-    final long sideOccupancy = bitboardPosition.occupied(side);
-    return sideOccupancy == (sideKings | sideKnights) && Long.bitCount(sideKings) == 1 && Long.bitCount(sideKnights) == 1;
-  }
-
-  static boolean calculateHasKingAndBishopsOnly(Side side, BitboardPosition bitboardPosition, SquareType squareType) {
-    final long sideKings = side == Side.WHITE ? bitboardPosition.whiteKings() : bitboardPosition.blackKings();
-    final long sideBishops = side == Side.WHITE ? bitboardPosition.whiteBishops() : bitboardPosition.blackBishops();
-    final long sideOccupancy = bitboardPosition.occupied(side);
-    if (sideOccupancy != (sideKings | sideBishops)) {
-      return false;
-    }
-    if (Long.bitCount(sideKings) != 1) {
-      return false;
-    }
-    final long colourMask = squareType == SquareType.LIGHT_SQUARE ? LIGHT_SQUARES : DARK_SQUARES;
-    return (sideBishops & colourMask) != 0L && (sideBishops & ~colourMask) == 0L;
-  }
-
-  private static int countBishops(Side side, BitboardPosition bitboardPosition, SquareType squareType) {
-    final long bishops = side == Side.WHITE ? bitboardPosition.whiteBishops() : bitboardPosition.blackBishops();
-    final long colourMask = squareType == SquareType.LIGHT_SQUARE ? LIGHT_SQUARES : DARK_SQUARES;
-    return Long.bitCount(bishops & colourMask);
-  }
-
-  // ===========================================================================================
-  // StaticPosition variants — legacy callers. To be removed once all sites switch to bitboard.
-  // ===========================================================================================
+  // --- private helpers ---
 
   private static boolean hasAnySide(PieceType pieceType, StaticPosition staticPosition) {
     return hasPieceType(Side.WHITE, pieceType, staticPosition) || hasPieceType(Side.BLACK, pieceType, staticPosition);
@@ -260,5 +163,4 @@ abstract class UnwinnabilityMaterial implements EnumConstants {
     }
     return total;
   }
-
 }
